Added custom converters to Configuration
The 3rd argument when accessing a value in a ConfigFile object can
now be a function instead of a type. The string value will be passed
to the provided function and what that returns will be the return of
the call.

Ex:
config file:
[section]
name=steven j vanhorn

code:
def listConverter(data):
    return data.split()

config["section","name",listConverter] # returns ["steven","j","vanhorn"]